backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "googlemaps>=4.10.0",
    "langchain-community>=0.3.29",
    "langchain-core>=0.3.76",
    "langchain-google-vertexai>=2.1.2",
    "langgraph>=0.6.7",
    "ortools>=9.14.6206",
    "python-dotenv>=1.1.1",
]
